[autogenerated] fix Inkscape SVG bug
Windows version of Inkscape *can* convert path with a Linux slash on the fly. 
Linux version of Inkscape *can't* convert path of Windows version backslash on the fly.
It result often with broken background image on the renderfarm.
This commit is autogenerated by a bash script doing the autofix before render.